Create 3rdPersonControler
I was able to interface with an inertial measurement unit to get the orientation of the robot. By using a PS4 Controller and moving the joystick in an intended direction, the robot will turn facing the direction that the stick is pointing in. (Assuming the controller is angled perpendicular to earth's magnetic poles & parallel with the ground. To expand the project, an IMU can be placed on the PS4 controller to eliminate the need to keep the controller fixed relative to the orientation of the earth.